backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
selenium==4.27.0
pillow==11.0.0
python-multipart==0.0.12
aiofiles==24.1.0
python-dotenv==1.0.1
